# openapi.yaml
openapi: '3.0.2'
info:
    title: Torrents Stream Server
    version: '1.0'
paths:
...
    /api/torrents:
        get:
            operationId: getTorrents
            tags:
                - torrents
            security:
                - apiKey: []
            responses:
                200:
                    description: OK
                    content:
                        application/json:
                            schema:
                                type: array
                                items:
                                    $ref: '#/components/schemas/TorrentModel'
                401:
                    $ref: '#/components/responses/401'
                403:
                    $ref: '#/components/responses/403'
        post:
            operationId: createTorrent
            tags:
                - torrents
            parameters:
                - in: query
                  name: torrent
                  description: magnet or torrent link
                  required: true
                  allowReserved: true
                  schema:
                      type: string
            security:
                - apiKey: []
            responses:
                200:
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/TorrentModel'
                400:
                    $ref: '#/components/responses/400'
                401:
                    $ref: '#/components/responses/401'
                403:
                    $ref: '#/components/responses/403'
    /api/torrents/{infoHash}:
        delete:
            operationId: deleteTorrent
            tags:
                - torrents
            parameters:
                - in: path
                  name: infoHash
                  description: torrent info hash
                  required: true
                  schema:
                      type: string
            security:
                - apiKey: []
            responses:
                200:
                    $ref: '#/components/responses/200'
        get:
            operationId: getTorrent
            tags:
                - torrents
            parameters:
                - in: path
                  name: infoHash
                  description: torrent info hash
                  required: true
                  schema:
                      type: string
            security:
                - apiKey: []
            responses:
                200:
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/TorrentModel'
                404:
                    $ref: '#/components/responses/404'
                401:
                    $ref: '#/components/responses/401'
                403:
                    $ref: '#/components/responses/403'
...
components:
...
    schemas:
...
        TorrentModel:
            type: object
            properties:
                link:
                    type: string
                infoHash:
                    type: string
                name:
                    type: string
                started:
                    type: number
                updated:
                    type: number
                files:
                    type: array
                    items:
                        $ref: '#/components/schemas/TorrentFileModel'
                downloaded:
                    type: number
                downloadSpeed:
                    type: number
                playlist:
                    type: string
                streamZip:
                    type: string
            required:
                [
                    'link',
                    'infoHash',
                    'name',
                    'started',
                    'updated',
                    'files',
                    'downloaded',
                    'downloadSpeed',
                    'playlist',
                    'streamZip',
                ]
        TorrentFileModel:
            type: object
            properties:
                name:
                    type: string
                path:
                    type: string
                type:
                    type: string
                length:
                    type: number
                stream:
                    type: string
                streamZip:
                    type: string
            required: ['name', 'path', 'length', 'type', 'stream', 'streamZip']
...
        ApiErrorModel:
            type: object
            properties:
                error:
                    type: string
            required: ['error']
...
    responses:
        200:
            description: Ok
...
        400:
            description: Bad Request
            content:
                application/json:
                    schema:
                        $ref: '#/components/schemas/ApiErrorModel'
        401:
            description: Unauthorized
            content:
                application/json:
                    schema:
                        $ref: '#/components/schemas/ApiErrorModel'
        403:
            description: Forbidden
            content:
                application/json:
                    schema:
                        $ref: '#/components/schemas/ApiErrorModel'
        404:
            description: Resource not found
            content:
                application/json:
                    schema:
                        $ref: '#/components/schemas/ApiErrorModel'
...
    securitySchemes:
        apiKey:
            type: http
            scheme: bearer
servers:
    - url: http://127.0.0.1:3000/
      description: Dev server